Added type casts to (unsigned int) in LB_Hash calls to make lint happy. Mixing ints and unsigned ints should be OK in C, so everything worked even without explicit type cast.

+++ packages/zoltan/src/parmetis/parmetis_jostle.c
@@ -564,7 +564,7 @@ static int LB_ParMetis_Jostle(
 
     for (i=0; i< num_obj; i++){
       /* insert hashed elements into hash table */
-      j = LB_Hash(&(global_ids[i*num_gid_entries]), num_gid_entries, num_obj);
+      j = LB_Hash(&(global_ids[i*num_gid_entries]), num_gid_entries, (unsigned int)num_obj);
       hash_nodes[i].next = hashtab[j];
       hashtab[j] = &hash_nodes[i];
     }
@@ -863,7 +863,7 @@ static int LB_ParMetis_Jostle(
   
     /* Insert nodes into hash table */
     for (i=0; i< nrecv; i++){
-      j = LB_Hash(hash_nodes[i].gid, num_gid_entries, nrecv);
+      j = LB_Hash(hash_nodes[i].gid, num_gid_entries, (unsigned int)nrecv);
       hash_nodes[i].next = hashtab[j];
       hashtab[j] = &hash_nodes[i];
       if (lb->Debug_Level >= LB_DEBUG_ALL) {
@@ -1547,7 +1547,7 @@ static int hash_lookup (LB *lb, struct LB_hash_node **hashtab, LB_ID_PTR key,
   int i;
   struct LB_hash_node *ptr;
 
-  i = LB_Hash(key, lb->Num_GID, n);
+  i = LB_Hash(key, lb->Num_GID, (unsigned int)n);
   for (ptr=hashtab[i]; ptr != NULL; ptr = ptr->next){
     if (LB_EQ_GID(lb, ptr->gid, key))
       return (ptr->gno);
